Seperate stories for link and button, move theme to config and make button type configurable
