+++ stories/button.stories.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+import { text, select } from '@storybook/addon-knobs';
+import { withInfo } from '@storybook/addon-info';
+import { Button } from '../src';
+
+const buttonStories = storiesOf('Button', module);
+
+buttonStories.add(
+  'configurable',
+  withInfo()(() => (
+    <Button type={select('Type', ['primary', 'secondary'], 'primary')}>{text('Children', 'Primary')}</Button>
+  ))
+);
